# Conflicts:
#	doc/modules/cassandra/pages/developing/cql/changes.adoc
#	pylib/cqlshlib/cql3handling.py
#	pylib/cqlshlib/test/test_cqlsh_completion.py
#	src/antlr/Parser.g
#	src/java/org/apache/cassandra/cql3/Operator.java
#	src/java/org/apache/cassandra/cql3/Relation.java
#	src/java/org/apache/cassandra/cql3/restrictions/SimpleRestriction.java
#	test/unit/org/apache/cassandra/cql3/restrictions/ClusteringElementsTest.java
#	test/unit/org/apache/cassandra/cql3/validation/operations/DeleteTest.java
#	test/unit/org/apache/cassandra/cql3/validation/operations/SelectSingleColumnRelationTest.java

+++ .build/run-python-dtests.sh
@@ -25,36 +25,60 @@
 #
 ################################
 
+[ $DEBUG ] && set -x
 
 # help
-if [ "$#" -lt 1 ] || [ "$#" -gt 2 ] || [ "$1" == "-h" ]; then
+if [ "$#" -lt 1 ] || [ "$1" == "-h" ]; then
     echo ""
-    echo "Usage: run-python-dtest.sh test_type [split_chunk|test_regexp]"
+    echo "Usage: $0 [-a|-t|-c|-j|-h]"
+    echo "   -a Test target type: dtest, dtest-latest, ..."
+    echo "   -t Test name regexp to run."
+    echo "   -c Chunk to run in the form X/Y: Run chunk X from a total of Y chunks."
     echo ""
     echo "        default split_chunk is 1/1"
     exit 1
 fi
 
 # Pass in target to run, defaults to dtest
-DTEST_TARGET="${1:-dtest}"
-# Optional: pass in chunk to test, formatted as "K/N" for the Kth chunk of N chunks
-DTEST_SPLIT_CHUNK="$2"
+DTEST_TARGET="dtest"
+
+# TODO implement repeated runs, eg CASSANDRA-18942
+while getopts "a:t:c:hj:" opt; do
+  case $opt in
+    a ) DTEST_TARGET="$OPTARG"
+        ;;
+    t ) DTEST_SPLIT_CHUNK="$OPTARG"
+        ;;
+    c ) DTEST_SPLIT_CHUNK="$OPTARG"
+        ;;
+    h ) print_help
+        exit 0
+        ;;
+    j ) ;; # To avoid failing on java_version param from docker/run_tests.sh
+    \?) error 1 "Invalid option: -$OPTARG"
+        ;;
+  esac
+done
+shift $((OPTIND-1))
+if [ "$#" -ne 0 ]; then
+  error 1 "Unexpected arguments"
+fi
 
 # variables, with defaults
 [ "x${CASSANDRA_DIR}" != "x" ] || CASSANDRA_DIR="$(readlink -f $(dirname "$0")/..)"
-[ "x${CASSANDRA_DTEST_DIR}" != "x" ] || CASSANDRA_DTEST_DIR="${CASSANDRA_DIR}/../cassandra-dtest"
+[ "x${CASSANDRA_DTEST_DIR}" != "x" ] || CASSANDRA_DTEST_DIR="$(readlink -f ${CASSANDRA_DIR}/../cassandra-dtest)"
 [ "x${DIST_DIR}" != "x" ] || DIST_DIR="${CASSANDRA_DIR}/build"
+[ "x${TMPDIR}" != "x" ] || { TMPDIR_SET=1 && export TMPDIR="$(mktemp -d ${DIST_DIR}/run-python-dtest.XXXXXX)" ; }
+[ "x${CCM_CONFIG_DIR}" != "x" ] && ls $CCM_CONFIG_DIR
 
-export TMPDIR="$(mktemp -d ${DIST_DIR}/run-python-dtest.XXXXXX)"
 export PYTHONIOENCODING="utf-8"
 export PYTHONUNBUFFERED=true
 export CASS_DRIVER_NO_EXTENSIONS=true
 export CASS_DRIVER_NO_CYTHON=true
 export CCM_MAX_HEAP_SIZE="1024M"
 export CCM_HEAP_NEWSIZE="512M"
-export CCM_CONFIG_DIR="${TMPDIR}/.ccm"
 export NUM_TOKENS="16"
-#Have Cassandra skip all fsyncs to improve test performance and reliability
+# Have Cassandra skip all fsyncs to improve test performance and reliability
 export CASSANDRA_SKIP_SYNC=true
 unset CASSANDRA_HOME
 
@@ -63,6 +87,8 @@ command -v ant >/dev/null 2>&1 || { echo >&2 "ant needs to be installed"; exit 1
 command -v virtualenv >/dev/null 2>&1 || { echo >&2 "virtualenv needs to be installed"; exit 1; }
 [ -f "${CASSANDRA_DIR}/build.xml" ] || { echo >&2 "${CASSANDRA_DIR}/build.xml must exist"; exit 1; }
 [ -d "${DIST_DIR}" ] || { mkdir -p "${DIST_DIR}" ; }
+ALLOWED_DTEST_VARIANTS="novnode|large|latest|upgrade"
+[[ "${DTEST_TARGET}" =~ ^dtest(-(${ALLOWED_DTEST_VARIANTS}))*$ ]] || { echo >&2 "Unknown dtest target: ${DTEST_TARGET}. Allowed variants are ${ALLOWED_DTEST_VARIANTS}"; exit 1; }
 
 java_version=$(java -version 2>&1 | awk -F '"' '/version/ {print $2}' | awk -F. '{print $1}')
 version=$(grep 'property\s*name=\"base.version\"' ${CASSANDRA_DIR}/build.xml |sed -ne 's/.*value=\"\([^"]*\)\".*/\1/p')
@@ -113,31 +139,26 @@ pip3 freeze
 cd ${CASSANDRA_DTEST_DIR}
 
 set +e # disable immediate exit from this point
-if [ "${DTEST_TARGET}" = "dtest" ]; then
-    DTEST_ARGS="--use-vnodes --num-tokens=${NUM_TOKENS} --skip-resource-intensive-tests"
-elif [ "${DTEST_TARGET}" = "dtest-novnode" ]; then
-    DTEST_ARGS="--skip-resource-intensive-tests --keep-failed-test-dir"
-elif [ "${DTEST_TARGET}" = "dtest-latest" ]; then
-    DTEST_ARGS="--use-vnodes --num-tokens=${NUM_TOKENS} --configuration-yaml=cassandra_latest.yaml --skip-resource-intensive-tests"
-elif [ "${DTEST_TARGET}" = "dtest-large" ]; then
-    DTEST_ARGS="--use-vnodes --num-tokens=${NUM_TOKENS} --only-resource-intensive-tests --force-resource-intensive-tests"
-elif [ "${DTEST_TARGET}" = "dtest-large-novnode" ]; then
-    DTEST_ARGS="--only-resource-intensive-tests --force-resource-intensive-tests"
-elif [ "${DTEST_TARGET}" = "dtest-upgrade" ]; then
-    DTEST_ARGS="--use-vnodes --num-tokens=${NUM_TOKENS} --execute-upgrade-tests --execute-upgrade-tests-only --upgrade-target-version-only --upgrade-version-selection all"
-elif [ "${DTEST_TARGET}" = "dtest-upgrade-novnode" ]; then
-    DTEST_ARGS="--execute-upgrade-tests --execute-upgrade-tests-only --upgrade-target-version-only --upgrade-version-selection all"
-elif [ "${DTEST_TARGET}" = "dtest-upgrade-large" ]; then
-    DTEST_ARGS="--use-vnodes --num-tokens=${NUM_TOKENS} --execute-upgrade-tests --execute-upgrade-tests-only --upgrade-target-version-only --upgrade-version-selection all --only-resource-intensive-tests --force-resource-intensive-tests"
-elif [ "${DTEST_TARGET}" = "dtest-upgrade-novnode-large" ]; then
-    DTEST_ARGS="--execute-upgrade-tests --execute-upgrade-tests-only --upgrade-target-version-only --upgrade-version-selection all --only-resource-intensive-tests --force-resource-intensive-tests"
+DTEST_ARGS="--keep-failed-test-dir"
+# Check for specific keywords in DTEST_TARGET and append corresponding options
+if [[ "${DTEST_TARGET}" == *"-large"* ]]; then
+    DTEST_ARGS+=" --only-resource-intensive-tests --force-resource-intensive-tests"
 else
-    echo "Unknown dtest target: ${DTEST_TARGET}"
-    exit 1
+    DTEST_ARGS+=" --skip-resource-intensive-tests"
+fi
+if [[ "${DTEST_TARGET}" != *"-novnode"* ]]; then
+    DTEST_ARGS+=" --use-vnodes --num-tokens=${NUM_TOKENS}"
+fi
+if [[ "${DTEST_TARGET}" == *"-latest"* ]]; then
+    DTEST_ARGS+=" --configuration-yaml=cassandra_latest.yaml"
+fi
+if [[ "${DTEST_TARGET}" == *"-upgrade"* ]]; then
+    DTEST_ARGS+=" --execute-upgrade-tests --execute-upgrade-tests-only --upgrade-target-version-only --upgrade-version-selection all"
 fi
 
 touch ${DIST_DIR}/test_list.txt
 ./run_dtests.py --cassandra-dir=${CASSANDRA_DIR} ${DTEST_ARGS} --dtest-print-tests-only --dtest-print-tests-output=${DIST_DIR}/test_list.txt 2>&1 > ${DIST_DIR}/test_stdout.txt
+
 [[ $? -eq 0 ]] || { cat ${DIST_DIR}/test_stdout.txt ; exit 1; }
 
 if [[ "${DTEST_SPLIT_CHUNK}" =~ ^[0-9]+/[0-9]+$ ]]; then
@@ -145,6 +166,12 @@ if [[ "${DTEST_SPLIT_CHUNK}" =~ ^[0-9]+/[0-9]+$ ]]; then
     ( split --help 2>&1 ) | grep -q "r/K/N" || split_cmd=gsplit
     command -v ${split_cmd} >/dev/null 2>&1 || { echo >&2 "${split_cmd} needs to be installed"; exit 1; }
     SPLIT_TESTS=$(${split_cmd} -n r/${DTEST_SPLIT_CHUNK} ${DIST_DIR}/test_list.txt)
+    if [[ -z "${SPLIT_TESTS}" ]]; then
+      # something has to run in the split to generate a nosetest xml result (and to not rerun all tests)
+      echo "Hacking ${DTEST_TARGET} to run only first test found as no tests in split ${DTEST_SPLIT_CHUNK} were found: "
+      SPLIT_TESTS="$( echo ${DIST_DIR}/test_list.txt | head -n1)"
+      echo "  ${SPLIT_TESTS}"
+    fi
     SPLIT_STRING="_${DTEST_SPLIT_CHUNK//\//_}"
 elif [[ "x" != "x${DTEST_SPLIT_CHUNK}" ]] ; then
     SPLIT_TESTS=$(grep -e "${DTEST_SPLIT_CHUNK}" ${DIST_DIR}/test_list.txt)
@@ -152,11 +179,16 @@ elif [[ "x" != "x${DTEST_SPLIT_CHUNK}" ]] ; then
 else
     SPLIT_TESTS=$(cat ${DIST_DIR}/test_list.txt)
 fi
+SPLIT_TESTS="${SPLIT_TESTS//$'\n'/ }"
 
 pytest_results_file="${DIST_DIR}/test/output/nosetests.xml"
 pytest_opts="-vv --log-cli-level=DEBUG --junit-xml=${pytest_results_file} --junit-prefix=${DTEST_TARGET} -s"
 
-pytest ${pytest_opts} --cassandra-dir=${CASSANDRA_DIR} --keep-failed-test-dir ${DTEST_ARGS} ${SPLIT_TESTS} 2>&1 | tee -a ${DIST_DIR}/test_stdout.txt
+echo ""
+echo "pytest ${pytest_opts} --cassandra-dir=${CASSANDRA_DIR} --keep-failed-test-dir ${DTEST_ARGS} ${SPLIT_TESTS}" 
+echo ""
+
+pytest ${pytest_opts}  --cassandra-dir=${CASSANDRA_DIR} --keep-failed-test-dir ${DTEST_ARGS} ${SPLIT_TESTS} 2>&1 | tee -a ${DIST_DIR}/test_stdout.txt
 
 # tar up any ccm logs for easy retrieval
 if ls ${TMPDIR}/*/test/*/logs/* &>/dev/null ; then
@@ -181,8 +213,10 @@ popd  >/dev/null
 #
 ################################
 
-[[ "${TMPDIR}" == *"${DIST_DIR}/run-python-dtest."* ]] && rm -rf "${TMPDIR}"
-unset TMPDIR
+if [ ${TMPDIR_SET} ] ; then
+    [[ "${TMPDIR}" == *"${DIST_DIR}/run-python-dtest."* ]] && rm -rf "${TMPDIR}"
+    unset TMPDIR
+fi
 deactivate
 
 # Exit cleanly for usable "Unstable" status
